Separate classes into java files
